Refactored overview cards
