+++ carbonmark-data/app/[locale]/details/price-of-digital-carbon/page.tsx
@@ -1,5 +1,5 @@
 import { t } from "@lingui/macro";
-import HistoricalPriceCard from "components/cards/HistoricalPriceCard";
+import HistoricalPriceCard from "components/cards/overview/HistoricalPriceCard";
 import TokensPriceCard from "components/cards/overview/TokensPriceCard";
 import DetailPage from "components/pages/DetailPage";
 import layout from "theme/layout.module.scss";
